Eliminate skonfig.cdist wrapper module
